requirements.txt
setuptools 
frozen-flask 
flask 
libsass 
markdown2 
python-slugify 
babel
python-frontmatter
pydantic